feat:: 스토리 삭제 시 서버에 저장된 이미지도 삭제하도록 수정
- 본인 스토리가 아닌경우 오류 반환

+++ src/story/story.service.ts
@@ -95,6 +95,11 @@ export class StoryService {
     userId: number,
     storyId: number,
   ): Promise<DeleteStoryResponseDto> {
+    const story = await this.storyRepository.findStoryById(storyId);
+    if (story.userId !== userId) {
+      throw new BadRequestException('Not your story!');
+    }
+    await this.fileService.deleteFile(story.imgDir);
     const isDeleted = await this.storyRepository.deleteStory(userId, storyId);
     if (!isDeleted) {
       throw new NotImplementedException('Delete story failed!');
